Some adds

+++ headers/IMatrix.h
@@ -10,8 +10,8 @@ class IMatrix:public IDrawable {
 public:
     virtual unsigned long getRows() const = 0;
     virtual unsigned long getCols() const = 0;
-    virtual int get(int, int) const = 0;
-    virtual void set(int, int, int) = 0;
+    virtual int get(unsigned long, unsigned long) const = 0;
+    virtual void set(unsigned long, unsigned long, int) = 0;
     virtual ~IMatrix() = default;
 };
 #endif //LAB1_2_IMATRIX_H
