+++ sources/VectorNormal.cpp
@@ -9,10 +9,10 @@ unsigned long VectorNormal::getSize() const {
     return this->data.size();
 }
 
-int VectorNormal::get(int pos) const {
+int VectorNormal::get(unsigned long pos) const {
     return this->data.at(pos);
 }
 
-void VectorNormal::set(int pos, int val) {
+void VectorNormal::set(unsigned long pos, int val) {
     this->data.at(pos) = val;
 }
